requirements.txt
psutil
pandas
PySide6
google-generativeai